pad_images must be in utils.py
