+++ terratorch/models/utils.py
@@ -1,3 +1,6 @@
+import torch
+from torch import nn, Tensor
+
 class DecoderNotFoundError(Exception):
     pass
 
@@ -11,3 +14,16 @@ def extract_prefix_keys(d: dict, prefix: str) -> dict:
             remaining_dict[k] = v
 
     return extracted_dict, remaining_dict
+
+def pad_images(imgs: Tensor,patch_size: int, padding:str) -> Tensor:
+    p = patch_size
+    # h, w = imgs.shape[3], imgs.shape[4]
+    t, h, w = imgs.shape[-3:]
+    h_pad, w_pad = (p - h % p) % p, (p - w % p) % p  # Ensure padding is within bounds
+    if h_pad > 0 or w_pad > 0:
+        imgs = torch.stack([
+            nn.functional.pad(img, (0, w_pad, 0, h_pad), mode=padding)
+            for img in imgs  # Apply per image to avoid NotImplementedError from torch.nn.functional.pad
+        ])
+    return imgs
+
